+++ CHANGELOG.md
@@ -2,8 +2,20 @@
 
 ## Unreleased
 
+### Added
+
+  - Hook - New non-standard cimi resource that provides an access
+    for events driven workflows
+  - Vendor - New resource allowing a user to create a Stripe 
+    Express account
+
 ### Changed
 
+  - Customer - Save subscription id in customer resource
+  - Deployment - New upcoming invoice operation
+  - Deployment - Subscription creation
+  - Module - New licence and price attributes
+  - Configuration - Nuvla get new stripe-client-id attribute
   - Nuvlabox status - New optional container-plugins attribute
   - Nuvlabox peripheral - schema accept peripheral telemetry
   - Deployment - Add check-dct action on deployment